Make use of a uint16_t

+++ src/ItemProgress.h
@@ -65,9 +65,9 @@ class ItemProgress : public MenuItem {
 
     /**
      * Set the progress on the item
-     * @param uint8_t progress for the item
+     * @param uint16_t progress for the item
      */
-    void setProgress(uint8_t value) override {
+    void setProgress(uint16_t value) override {
         if (progress < MIN_PROGRESS || progress > MAX_PROGRESS) return;
         progress = value;
     }
